Bigtable input available for training
Ability to sample moves from last N games in table.
This is a preview, for discussion and for experimenting;
right now it fixes N=500k, and the probability; these
should be parameters.

Google Cloud Bigtable package now required.

In the C++ Bigtable code:
  - Use 3 digits for move number, and batch game moves
  - Factor out row patterns as constants.

Also, skip lint of `bigtable_input.py`, since it's still a work in
progress.