Add tfrzz_to_cbt game conversion tool
Converts Minigo game examples from the `.tfrecord.zz`
format in GCS, in batch, to a Bigtable target.
Uses configurable multiprocessing to increase the rate
and reset memory growth in the Bigtable C++ client.

Can specify GCS file batches with `--glob_pattern`,
and also specify input files as positional arguments.

It converts synchronously by default, blocking the caller
until all the games have been converted, then reports on
rate statistics.  When aysynchronous, the tool forks
and returns as soon as it has updated the game counter,
which makes it friendly to any fleet orchestration:
as soon as the tool exits, the next conversion can be
undertaken on any machine, anywhere, since the game
counter is atomically stored in the table itself.