Remove batch_size flag.
Make it easier to create a DualNet without wrapping it in a
BatchingDualNet.

Dynamically resize DualNets so that we can contain the batch size within
BatchingDualNet.

The DualNet interface is exposing the buffer count, which is really all
that's needed to determine the (theoretically) optimal batch size.

Might trigger reallocations, and the TrtDualNet has a hard (coded) batch
size limit. We might want to expose a Reserve() function on the DualNet
interface so that the BatchingDualNet can call it when it determined the
batch size.