Store move count per game for precise selection
In order to count moves in a range of games quickly, store the sum of
moves per game in a new `move_count` column.  They are stored both in
the zero-move row for the game, but also in a new `ct_` key space,
which can be faster to scan.  Using this new column, the exact number
of moves in 500e3 games can be calculated in 2-8 seconds.

By default, use a performant cut-and-shuffle technique for the moves
selected from the game range: randomly interleave multiple shards and
then shuffle the result with a buffer 2X the buffer size chosen from
the shards.  This gives reasonably good shuffling with only a 20%
performance hit over not shuffling.

Also fixed:

  - Off-by-one error in game range
  - Row keys must be `bytes`, not `str`