During self-play, write bleakest move to Bigtable (#552)
* During self-play, write bleakest move to Bigtable

The Q value will be written to a column named
`metadata:bleakest_q` for the move row at which the
evaluation occurred, for games in which resigning
has been disabled (bleakest-move calculation is
not relevant otherwise, since quitters don't know
how bad it can get).  This makes it straightforward
to collect the games, moves, and Q values for
bleakest-move evaluations over a range of games,
e.g.:

  `cbt read games columns=metadata:bleakest_q`

The only rows selected will be those with a value
in the column, and the row name will contain the
game and move number of the evaluation.

Also reduced some coupling and refactored the
calculation of bleakest move to a function callable
from more than `main.cc`.