Adapt minigo python engine to use EdgeTPU by adding a custom dual_net.py.
This supports both the EdgeTPU dev board and the USB form factor for
Linux and Raspberry Pi hosts. In order not to use dependencies that
aren't present or available on all plattforms the imports are inside
the if statement. Potentially this could be moved to it's own
dual_net_chooser.py or something.